+++ cnccode/business_classes/BUActivity.inc.php
@@ -43,6 +43,7 @@ require_once($cfg ["path_bu"] . "/BUStandardText.inc.php");
 require_once($cfg["path_dbe"] . "/DBEJPorhead.inc.php");
 require_once($cfg["path_dbe"] . "/DBEPendingReopened.php");
 require_once($cfg["path_ct"] . "/CTProject.inc.php");
+require_once($cfg ["path_bu"] . "/BUProblemRaiseType.inc.php");
 
 define(
     'BUACTIVITY_RESOLVED',
@@ -5307,7 +5308,6 @@ is currently a balance of ';
                 $dbeCustomerItem->getPKValue()
             );
             $this->dbeProblem->updateRow();
-
         } else {
             $this->raiseError('No Pre-pay Contract Found');
             return FALSE;
@@ -5423,6 +5423,7 @@ is currently a balance of ';
         );
 
         $dbeCallActivity->insertRow();
+        $this->setProblemRaise($dbeProblem,$dbeCallActivity); //createActivityFromCustomerID
 
         return $dbeCallActivity->getPKValue();
     }
@@ -5450,6 +5451,7 @@ is currently a balance of ';
         /*
     * Create a new problem
     */
+    
         $dbeProblem = new DBEProblem($this);
         $dbeProblem->setValue(
             DBEProblem::hdLimitMinutes,
@@ -5543,6 +5545,9 @@ is currently a balance of ';
             DBEProblem::projectID,
             @$_SESSION [$sessionKey] ['projectID']
         );
+        
+        
+
         $dbeProblem->insertRow();
 
         if ($_SESSION[$sessionKey]['monitorSRFlag'] === 'Y') {
@@ -5625,6 +5630,8 @@ is currently a balance of ';
             $GLOBALS['auth']->is_authenticated()
         ); // user that created activity
         $dsCallActivity->post();
+        $this->setProblemRaise($dbeProblem,$dsCallActivity); //createActivityFromSession
+
         $dbeContact = null;
         if (@$_SESSION[$sessionKey]['contactID']) {
             $dbeContact = new DBEContact($this);
@@ -5666,6 +5673,88 @@ is currently a balance of ';
         unset($_SESSION[$sessionKey]);
 
         return $dsCallActivity;
+    }
+    private function getProblemRaiseType($description)
+    {      
+        $buProblemRaiseType =new BUProblemRaiseType($this);
+        $id=null;
+        $id= $buProblemRaiseType->getProblemRaiseTypeByName($description)['id'];              
+        return $id;
+    }
+    private function setProblemRaise($dbeProblem, $callActivity, $raiseType = null)
+    {
+        if (!isset($dbeProblem) && !isset($callActivity))
+            return null;
+        if ($raiseType != null) {
+            $dbeProblem->setValue(
+                DBEProblem::problemraisetypeId,
+                $this->getProblemRaiseType($raiseType)
+            );
+            $dbeProblem->updateRow();
+            return;
+        }
+        if (isset($GLOBALS['auth'])) {
+            // get team
+            $userID = $GLOBALS['auth']->is_authenticated();
+            if (isset($userID)) {
+                $dbeUser = new DBEUser($this);
+                $dbeUser->setPKValue($userID);
+                $dbeUser->getRow();
+                $teamId = $dbeUser->getValue(DBEUser::teamID);
+            }
+        }
+      
+        
+        //For each problem, where callactivity.caa_callacttypeno = 57 and callactivity.caa_serverguard = Y, then set the problem source as Alert.
+          if (
+            isset($dbeProblem) && isset($callActivity)
+            && $callActivity->getValue(DBEJCallActivity::callActTypeID) == 57
+            && $callActivity->getValue(DBEJCallActivity::serverGuard) == 'Y'
+        ) {
+            $dbeProblem->setValue(
+                DBEProblem::problemraisetypeId,
+                $this->getProblemRaiseType(BUProblemRaiseType::ALERT)
+            );
+        } else if (
+            isset($dbeProblem) && isset($callActivity)
+            && $dbeProblem->getValue(DBEJProblem::linkedSalesOrderID) > 0
+            //&& $dbeProblem->getValue(DBEJProblem::priority) == 5
+        ) {
+            $dbeProblem->setValue(
+                DBEProblem::problemraisetypeId,
+                $this->getProblemRaiseType(BUProblemRaiseType::SALES)
+            );
+        } else if //For each problem, where callactivity.caa_callactivityno = 57 and callactivity.caa_consno = 67 and callactivity.caa_serverguard = N, then set the problem source as Email.
+        (
+            isset($dbeProblem) && isset($callActivity)
+            && $callActivity->getValue(DBEJCallActivity::callActTypeID) == 57
+            && $callActivity->getValue(DBEJCallActivity::caaConsno) == 67
+            && $callActivity->getValue(DBEJCallActivity::serverGuard) == 'N'
+        ) {
+            $dbeProblem->setValue(
+                DBEProblem::problemraisetypeId,
+                $this->getProblemRaiseType(BUProblemRaiseType::EMAIL)
+            );
+        } else if (isset($teamId) && $teamId == 1) //created by help desk
+        {
+            $raiseType=BUProblemRaiseType::PHONE;
+            if($dbeUser->getValue(DBEUser::basedAtCustomerSite)==1 &&
+            $dbeProblem->getValue(DBEProblem::customerID) == $dbeUser->getValue(DBEUser::siteCustId)  )
+                $raiseType=BUProblemRaiseType::ONSITE;
+            $dbeProblem->setValue(
+                DBEProblem::problemraisetypeId,
+                $this->getProblemRaiseType($raiseType)
+            );
+        }
+        else {
+            $dbeProblem->setValue(
+                DBEProblem::problemraisetypeId,
+                $this->getProblemRaiseType(BUProblemRaiseType::MANUAL)
+            );
+        }
+
+        if (isset($dbeProblem))
+            $dbeProblem->updateRow();
     }
 
     public function toggleMonitoringFlag($problemID)
@@ -6876,6 +6965,8 @@ is currently a balance of ';
             }
         }
 
+             
+       
         $dbeProblem->insertRow();
 
         $reason = "<p>An order has been received for the items below:</p>";
@@ -6999,9 +7090,8 @@ is currently a balance of ';
             );
         }
         //$dbeCallActivity->setValue( 'overtimeExportedFlag', 'N' );
-
         $dbeCallActivity->insertRow();
-
+        $this->setProblemRaise($dbeProblem,$dbeCallActivity,BUProblemRaiseType::SALES); //createSalesServiceRequest
         $db = new dbSweetcode(); // database connection for query
 
         $sql =
@@ -7351,7 +7441,6 @@ FROM
             $dbeProblem = new DBEProblem($this);
 
             $dbeProblem->getRow($automatedRequest->getServiceRequestID());
-
             if (!$dbeProblem->rowCount()) {
                 echo "<div>The service request doesn't exist </div>";
                 // create a new service request
@@ -7801,6 +7890,12 @@ FROM
             DBEJProblem::userID,
             null
         );        // not allocated
+        
+        $raiseTypeId= $record->getServerGuardFlag()=='Y'?BUProblemRaiseType::ALERTID: BUProblemRaiseType::EMAILID;
+        $dbeProblem->setValue(
+            DBEJProblem::problemraisetypeId,
+            $raiseTypeId
+        );  
         $dbeProblem->insertRow();
 
 
@@ -7867,7 +7962,8 @@ FROM
         );
 
         $dbeCallActivity->insertRow();
-
+       // $this->setProblemRaise($dbeProblem,$dbeCallActivity); // raiseNewRequestFromImport
+        
         if ($record->getAttachment() == 'Y') {
             $this->processAttachment(
                 $dbeProblem->getPKValue(),
@@ -9804,6 +9900,10 @@ FROM
                 DBEProblem::userID,
                 null
             );        // not allocated
+            $dbeProblem->setValue(
+                DBEProblem::problemraisetypeId,
+                BUProblemRaiseType::ALERTID
+            ); 
             $dbeProblem->insertRow();
 
             $problemID = $dbeProblem->getPKValue();
@@ -9855,7 +9955,7 @@ FROM
             );
 
             $dbeCallActivity->insertRow();
-
+            
         } else {
             $this->createFollowOnActivity(
                 $callActivityID,
@@ -9946,7 +10046,8 @@ FROM
             $detailsWithoutDriveLetters,
             $details,
             $serverName,
-            $serverCustomerItemID
+            $serverCustomerItemID,
+            BUProblemRaiseType::ALERTID
         );
     }
 
@@ -9965,7 +10066,8 @@ FROM
         $matchText,
         $details,
         $serverName,
-        $serverCustomerItemID
+        $serverCustomerItemID,
+        $raiseTypeId=null
     )
     {
         $priority = 2;
@@ -10058,6 +10160,11 @@ FROM
                 DBEProblem::userID,
                 null
             );        // not allocated
+            if($raiseTypeId!=null)
+                $dbeProblem->setValue(
+                    DBEProblem::problemraisetypeId,
+                    $raiseTypeId
+                ); 
             $dbeProblem->insertRow();
 
             $problemID = $dbeProblem->getPKValue();
@@ -10120,8 +10227,7 @@ FROM
             );
 
             $dbeCallActivity->insertRow();
-
-        } else {
+         } else {
 
             $this->createFollowOnActivity(
                 $callActivityID,
@@ -10161,7 +10267,8 @@ FROM
             $details,
             $details,
             $serverName,
-            $serverCustomerItemID
+            $serverCustomerItemID,
+            BUProblemRaiseType::ALERTID
         );
     }
 
@@ -10910,6 +11017,7 @@ FROM
                 )
             );
             $dbeCallActivity->insertRow();
+            $this->setProblemRaise($dbeProblem,$dbeCallActivity,BUProblemRaiseType::SALES); //sendSalesRequest
         }
 
         $buStandardText = new BUStandardText($this);