+++ cnccode/controller_classes/CTCNC.inc.php
@@ -663,6 +663,11 @@ class CTCNC extends Controller
                 "href"  => "SLAPerformance.php"
             ],
             [
+                "id"    => 224,
+                "label" => "SR Source",
+                "href"  => "SRSrource.php"
+            ],
+            [
                 "id"    => 209,
                 "label" => "3CX Call Reporting",
                 "href"  => "CallReporting.php"