+++ cnccode/controller_classes/CTUser.inc.php
@@ -342,7 +342,15 @@ class CTUser extends CTCNC
                 true
             );
         }
-
+        $siteCustomerString='';
+        $siteCustomerId=$dsUser->getValue(DBEJUser::siteCustId);
+        if(isset($siteCustomerId))
+        {
+            $dbeCustomer=new DBECustomer($this);
+            $dbeCustomer->setPKValue( $siteCustomerId);
+            $dbeCustomer->getRow();
+            $siteCustomerString=$dbeCustomer->getValue(DBECustomer::name);
+        }
 
         $this->template->setVar(
             array(
@@ -601,7 +609,9 @@ class CTUser extends CTCNC
                 'urlDelete'                                     => $urlDelete,
                 'txtDelete'                                     => $txtDelete,
                 'urlDisplayList'                                => $urlDisplayList,
-
+                "basedAtCustomerSiteChecked"                    => $this->dsUser->getValue(DBEUser::basedAtCustomerSite) ? 'checked' : null,
+                "basedAtCustomerSiteChecked"                    => $this->dsUser->getValue(DBEUser::basedAtCustomerSite) ? 'checked' : null,
+                'siteCustomerString'                            => $siteCustomerString,
             )
         );
         // manager selection
@@ -731,8 +741,10 @@ class CTUser extends CTCNC
      */
     function update()
     {
-        $this->setMethodName('update');
+        $this->setMethodName('update'); 
+              
         $this->formError = (!$this->dsUser->populateFromArray($this->getParam('user')));
+        
         $userData = $this->getParam('user')[1];
         $this->updateEncryptedData($userData, $this->dsUser);